server/requirements.txt
supabase==1.0.0
nltk==3.8.1
flask==2.2.3
flask_restful==0.3.9
flask_cors==3.0.10
pyOpenSSL==23.0.0